refactor tests to use async syntax
